Move memory module to coresight
