+++ probe-rs/src/cores/m4.rs
@@ -1,4 +1,4 @@
-use crate::memory::MI;
+use crate::coresight::memory::MI;
 use crate::probe::{DebugProbeError, MasterProbe};
 use crate::target::{
     BasicRegisterAddresses, Core, CoreInformation, CoreRegister, CoreRegisterAddress,
