+++ probe-rs/src/coresight/mod.rs
@@ -3,3 +3,4 @@ pub mod ap_access;
 pub mod common;
 pub mod debug_port;
 pub mod dp_access;
+pub mod memory;
