+++ probe-rs/src/lib.rs
@@ -16,7 +16,6 @@ pub mod cores;
 pub mod coresight;
 pub mod debug;
 pub mod flash;
-pub mod memory;
 pub mod probe;
 pub mod session;
 pub mod target;
